+++ spec/watir/web_scraper/scraper_spec.rb
@@ -6,29 +6,44 @@ RSpec.describe Watir::WebScraper::Scraper do
   let(:args) do
     {
       actions: actions,
-      params: params
+      params: params,
+      browser: browser
     }
   end
 
-  let(:actions) { [step_class, path_class, fetch_class] }
-
-  let(:step_class) { class_double(Watir::WebScraper::Step) }
-  let(:step_instance) { instance_double(Watir::WebScraper::Step) }
-
-  let(:path_class) { class_double(Watir::WebScraper::Path) }
-  let(:path_instance) { instance_double(Watir::WebScraper::Path) }
-
-  let(:fetch_class) { class_double(Watir::WebScraper::Fetcher) }
-  let(:fetch_instance) { instance_double(Watir::WebScraper::Fetcher) }
+  let(:browser) { class_double(Watir::WebScraper::Browser::Chrome) }
 
   let(:params) { {} }
 
+  before do
+    allow(browser).to receive(:start)
+    allow(browser).to receive(:new)
+  end
+
   describe '#perform' do
+    let(:actions) { [step_class, path_class, fetcher_class] }
+
+    let(:step_class) { class_double(Watir::WebScraper::Step) }
+    let(:step_instance) { instance_double(Watir::WebScraper::Step) }
+
+    let(:path_class) { class_double(Watir::WebScraper::Path) }
+    let(:path_instance) { instance_double(Watir::WebScraper::Path) }
+
+    let(:fetcher_class) { class_double(Watir::WebScraper::Fetcher) }
+    let(:fetcher_instance) { instance_double(Watir::WebScraper::Fetcher) }
+
     before do
-      [:step, :path, :fetch].each do |action|
-        allow(send("#{action}_class".to_sym)).to receive(:new).and_return(send("#{action}_instance".to_sym))
-        allow(send("#{action}_instance".to_sym)).to receive(:start)
-      end
+      allow(step_class).to receive(:new).and_return(step_instance)
+      allow(step_instance).to receive(:start)
+      allow(step_instance).to receive(:fetcher?).and_return(false)
+
+      allow(path_class).to receive(:new).and_return(path_instance)
+      allow(path_instance).to receive(:start)
+      allow(path_instance).to receive(:fetcher?).and_return(false)
+
+      allow(fetcher_class).to receive(:new).and_return(fetcher_instance)
+      allow(fetcher_instance).to receive(:start)
+      allow(fetcher_instance).to receive(:fetcher?).and_return(true)
     end
 
     context 'when routine is successful' do
@@ -82,6 +97,75 @@ RSpec.describe Watir::WebScraper::Scraper do
   describe '#actions' do
     subject { scraper.actions }
 
+    let(:actions) { [step_class] }
+    let(:step_class) { class_double(Watir::WebScraper::Step) }
+
     it { is_expected.to eq(actions) }
+  end
+
+  describe '#fetched_data' do
+    subject(:fetched_data) { scraper.fetched_data }
+
+    let(:actions) { [] }
+
+    it { is_expected.to be_a(Hash) }
+
+    context 'when fetcher action exists during perform routine' do
+      let(:actions) { [fetcher_class] }
+
+      let(:fetcher_class) { class_double(Watir::WebScraper::Fetcher) }
+      let(:fetcher) { instance_double(Watir::WebScraper::Fetcher) }
+      let(:test_data) { { test: :data } }
+
+      before do
+        allow(fetcher_class).to receive(:new).and_return(fetcher)
+        allow(fetcher).to receive(:fetcher?).and_return(true)
+        allow(fetcher).to receive(:start).and_return(fetcher)
+        allow(fetcher).to receive(:fetched_data).and_return(test_data)
+
+        scraper.perform
+      end
+
+      it 'returns data fetched by fetcher' do
+        expect(fetched_data).to eq(test_data)
+      end
+    end
+
+    context 'when multiple fetcher actions exist during perform routine' do
+      let(:actions) { [first_fetcher_class, second_fetcher_class] }
+
+      let(:first_fetcher_class) { class_double(Watir::WebScraper::Fetcher) }
+      let(:second_fetcher_class) { class_double(Watir::WebScraper::Fetcher) }
+
+      let(:first_fetcher) { instance_double(Watir::WebScraper::Fetcher) }
+      let(:second_fetcher) { instance_double(Watir::WebScraper::Fetcher) }
+
+      let(:expected_data) do
+        {
+          foo: :bar,
+          baz: :qux
+        }
+      end
+
+      before do
+        allow(first_fetcher_class).to receive(:new).and_return(first_fetcher)
+        allow(second_fetcher_class).to receive(:new).and_return(second_fetcher)
+
+        allow(first_fetcher).to receive(:start).and_return(first_fetcher)
+        allow(second_fetcher).to receive(:start).and_return(second_fetcher)
+
+        allow(first_fetcher).to receive(:fetcher?).and_return(true)
+        allow(second_fetcher).to receive(:fetcher?).and_return(true)
+
+        allow(first_fetcher).to receive(:fetched_data).and_return(foo: :bar)
+        allow(second_fetcher).to receive(:fetched_data).and_return(baz: :qux)
+
+        scraper.perform
+      end
+
+      it 'returns merged data fetched by all fetchers' do
+        expect(fetched_data).to eq(expected_data)
+      end
+    end
   end
 end