+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * [#13](https://github.com/CustomComm/watir-web_scraper/pull/13): Renames Fetcher#options to Fetcher#params
+* [#14](https://github.com/CustomComm/watir-web_scraper/pull/14): Support for multiple fetcher actions
 
 ### Bugfixes